+++ bin/miden-cli/src/commands/account.rs
@@ -125,8 +125,14 @@ impl AccountCmd {
                         println!("Current default account ID: {default_account}");
                     },
                     Some(id) if id == "none" => {
-                        client.remove_setting(DEFAULT_ACCOUNT_ID_KEY.to_string()).await?;
-                        println!("Removing default account...");
+                        let removed =
+                            client.remove_setting(DEFAULT_ACCOUNT_ID_KEY.to_string()).await?;
+
+                        if removed {
+                            println!("Default account removed");
+                        } else {
+                            println!("No default account was set");
+                        }
                     },
                     Some(id) => {
                         let account_id: AccountId = parse_account_id(&client, id).await?;