+++ bin/miden-cli/src/commands/address.rs
@@ -166,9 +166,12 @@ async fn remove_address<AUTH>(
     let address = decode_account_address(&address, account_id, &network_id)?;
     let note_tag = address.to_note_tag();
 
-    println!("removing address - Account Id {account_id} - Note tag: {note_tag}");
+    if client.remove_address(address, account_id).await? {
+        println!("Address removed - Account Id {account_id} - Note tag: {note_tag}");
+    } else {
+        println!("No address was tracked for Account Id {account_id}");
+    }
 
-    client.remove_address(address, account_id).await?;
     Ok(())
 }